feat(contact): SEO-focused hero text via data file, add location line, improve trust indicator alts, and scroll-to-form UX
